+++ anki-addon/ui/__init__.py
@@ -6,6 +6,7 @@ from .about_dialog import AboutDialog
 from .dashboard import DashboardDialog
 from .deck_rules_dialog import DeckRulesDialog
 from .discord_settings import DiscordSettingsDialog
+from .help_dialog import HelpDialog
 from .menu import ToolkitMenuManager, menu_manager
 from .templates_dialog import TemplatesDialog
 from .theme_settings import ThemeSettingsDialog
@@ -18,5 +19,6 @@ __all__ = [
     "DiscordSettingsDialog",
     "DeckRulesDialog",
     "TemplatesDialog",
+    "HelpDialog",
     "AboutDialog",
 ]